Apply CONTRIBUTING.md to all repositories
CONTRIBUTING.md goes to all repositories, but the GH PR template
only goes to OpenShift image repositories.

Refactor sync.sh to determine whether it is operating on an OpenShift
source or not, and act accordingly. This can be extended to perform
operations specific to not-OpenShift repositories in the future.

Update sync-all.sh (really, simplify) to simply run sync.sh for all
repositories and not perform any "if openshift" filtering itself.